% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_select_K_local}
\alias{cv_select_K_local}
\title{Cross-validated selection of K}
\usage{
cv_select_K_local(Y, X, K_grid, L, h)
}
\arguments{
\item{Y}{numeric vector}

\item{X}{numeric matrix/data.frame}

\item{K_grid}{integer vector of candidate K values}

\item{L}{number of folds}

\item{h}{bandwidth}
}
\value{
integer K_opt; attributes 'avg_MSE' and 'MSE_mat' are attached
}
\description{
Cross-validated selection of K
}
