% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{local_quantile}
\alias{local_quantile}
\title{Local quantile estimator at a point}
\usage{
local_quantile(Y, X, x, tau, h)
}
\arguments{
\item{Y}{numeric vector, response}

\item{X}{numeric matrix/data.frame, predictors (n x d)}

\item{x}{numeric vector of length d, target point}

\item{tau}{quantile level in (0,1)}

\item{h}{bandwidth (scalar or length-d vector; here treated as scalar)}
}
\value{
numeric scalar, estimated conditional tau-quantile at x
}
\description{
Local quantile estimator at a point
}
