% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{local_quantile_vec}
\alias{local_quantile_vec}
\title{Vectorized local quantile for multiple target points}
\usage{
local_quantile_vec(Y, X, X0, tau, h)
}
\description{
Vectorized local quantile for multiple target points
}
